[refactoring] modify directory tree
