+++ lib/harmonics_helper.rb
@@ -1,15 +1,18 @@
 require "harmonics_helper/version"
 
 require "harmonics_helper/xml_parser"
-require "harmonics_helper/single_part"
-require "harmonics_helper/pair_parts"
 require "harmonics_helper/prohibit_checker"
 
 require "harmonics_helper/output_formatter"
 
+require "harmonics_helper/part/single_part"
+require "harmonics_helper/part/pair_parts"
+require "harmonics_helper/part/four_parts"
+
 require "harmonics_helper/code/rotations"
 
 require "harmonics_helper/etc/exceptions"
+require "harmonics_helper/etc/validators"
 require "harmonics_helper/etc/path_helper"
 
 module HarmonicsHelper
