--- lib/harmonics_helper/harmony.rb
@@ -1,1 +1,0 @@
-# パート間の縦の関係を扱う
