--- lib/harmonics_helper/output.rb
@@ -1,1 +1,0 @@
-# 出力担当、ターミナルとあとテキストも欲しい
