+++ spec/pair_part_spec.rb
@@ -1,24 +1,24 @@
-RSpec.describe HarmonicsHelper::PairParts do
+RSpec.describe HarmonicsHelper::Part::PairParts do
 
   # These numbers comes from soprano and bass parts of MusicXMLSample
   SOPRANO_SOUNDS = [64, 65, 64, 62, 60, 65, 62, 60]
   SOPRANO_DURATION = [2, 2, 2, 2, 2, 2, 4, 4]
   BASS_SOUNDS = [36, 41, 36, 43, 45, 41, 43, 36] 
   BASS_DURATION = [2, 2, 2, 2, 2, 2, 4, 4]
-  let(:pair_parts) { HarmonicsHelper::PairParts.new(SOPRANO_SOUNDS, SOPRANO_DURATION, BASS_SOUNDS, BASS_DURATION)}
+  let(:pair_parts) { HarmonicsHelper::Part::PairParts.new(SOPRANO_SOUNDS, SOPRANO_DURATION, BASS_SOUNDS, BASS_DURATION)}
 
   describe "#initialize" do
     context "raise exception" do
       it "if all length is different raise Error" do
-        expect{ HarmonicsHelper::PairParts.new(SOPRANO_SOUNDS, SOPRANO_DURATION, BASS_SOUNDS, [2, 2, 2, 2, 2, 2, 2, 2]) }.to raise_error(StandardError)
+        expect{ HarmonicsHelper::Part::PairParts.new(SOPRANO_SOUNDS, SOPRANO_DURATION, BASS_SOUNDS, [2, 2, 2, 2, 2, 2, 2, 2]) }.to raise_error(StandardError)
       end
 
       it "if base args length is different raise Error" do
-        expect{ HarmonicsHelper::PairParts.new(SOPRANO_SOUNDS, SOPRANO_DURATION, BASS_SOUNDS, [2, 2, 2, 2, 2, 2, 2]) }.to raise_error(StandardError)
+        expect{ HarmonicsHelper::Part::PairParts.new(SOPRANO_SOUNDS, SOPRANO_DURATION, BASS_SOUNDS, [2, 2, 2, 2, 2, 2, 2]) }.to raise_error(StandardError)
       end
 
       it "if soprano args length is different raise Error" do
-        expect{ HarmonicsHelper::PairParts.new(SOPRANO_SOUNDS, [2, 2, 2, 2, 2, 2], BASS_SOUNDS, BASS_DURATION) }.to raise_error(StandardError)
+        expect{ HarmonicsHelper::Part::PairParts.new(SOPRANO_SOUNDS, [2, 2, 2, 2, 2, 2], BASS_SOUNDS, BASS_DURATION) }.to raise_error(StandardError)
       end
 
     end
