+++ spec/xml_parser_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe HarmonicsHelper::Parser do
 
   describe "#parts" do
     it "sample parts is [1,2,5,6]" do
-      expect(parser.parts).to eq([1,2,5,6])
+      expect(parser.parts).to eq([1, 2, 5, 6])
     end
   end
 
@@ -35,35 +35,6 @@ RSpec.describe HarmonicsHelper::Parser do
     end
   end
 
-  describe "#part_info" do
-    # sample soprano is 1 and bass is 6
-    it "sample soprano sound start with 5E" do
-      # 5E is 12*5 + 4 = 64
-      expect(parser.part_info(1)[0]["sound"]).to eq(64)
-    end
-  
-    it "sample soprano duration start with 2" do
-      expect(parser.part_info(1)[0]["duration"]).to eq(2)
-    end
-
-    it "sample soprano length is 8" do
-      expect(parser.part_info(1).length).to eq(8)
-    end
-
-    it "sample soprano length equal sample bass length" do
-      expect(parser.part_info(1).length).to eq(parser.part_info(6).length)
-    end
-
-    it "sample soprano is end with 5C" do
-      # 5C is 12*5 + 0 = 60
-      expect(parser.part_info(1)[7]["sound"]).to eq(60) 
-    end
-
-    it "sample soprano duration is end 4" do
-      expect(parser.part_info(1)[7]["duration"]).to eq(4) 
-    end
-  end
-
   describe "#beat" do
     it "sample beat is 4" do
       expect(parser.beat).to eq(4)
@@ -82,7 +53,7 @@ RSpec.describe HarmonicsHelper::Parser do
     end
 
     it "sample soprano durations[2,2,2,2,2,2,4,4]" do
-      expect(parser.durations(1)).to eq([2,2,2,2,2,2,4,4])
+      expect(parser.durations(1)).to eq([2, 2, 2, 2, 2, 2, 4, 4])
     end
   end
